pyproject.toml
[tool.poetry]
name = "godown.dao-backend"
version = "0.1.0-alpha"
description = "FastAPI project generated using minimal-fastapi-postgres-template."
authors = ["admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.29"}
python-dotenv = "^0.19.2"
pydantic = {extras = ["email"], version = "^1.9.0"}
alembic = "^1.7.5"
asyncpg = "^0.25.0"
fastapi-users = {extras = ["sqlalchemy2"], version = "^9.2.0"}
greenlet = "^1.1.2"
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
thefuzz = "^0.19.0"
gunicorn = "^20.1.0"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}
autoflake = "^1.4"
uvicorn = "^0.16.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
requests = "^2.27.1"
coverage = "^6.2"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
httpx = "^0.21.3"
sqlalchemy2-stubs = "^0.0.2-alpha.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v"
testpaths = [
    "app/tests",
]

[tool.isort]
profile = "black"